requirements.txt
google-cloud-storage
tqdm
requests
